+++ src/components/utils/UniqueComponentId.js
@@ -1,6 +1,6 @@
 var lastId = 0;
 
-export default function (prefix = 'i_id_') {
-    lastId++;
-    return `${prefix}${lastId}`;
+export default function (prefix = 'i-uid-') {
+  lastId++;
+  return `${prefix}${lastId}`;
 }