+++ src/components/utils/validator.ts
@@ -1,0 +1,2 @@
+export const isValidSize = (val: string) =>
+['', 'xl', 'lg', 'md', 'sm', 'xs'].includes(val)