+++ vue.config.js
@@ -1,6 +1,7 @@
 module.exports = {
   publicPath: process.env.NODE_ENV === 'production' ? './' : './',
   productionSourceMap: false,
+  lintOnSave: false,
   configureWebpack: {
     output: {
       libraryExport: 'default',